restructure code and Rename file, methods to be more meaniful
