+++ src/index.js
@@ -1,19 +1,17 @@
 import "./style.css";
-import displayInHtml, { getData, getSingeMeal,cardContainer } from "./modules/GetMeal.js";
+import { getSingeMeal, getAllMeal } from "./modules/getMeals.js";
+import { displayInHtml, cardContainer } from "./modules/displayMeals.js";
 import generatePopupMarkup from "./modules/generatePopupMarkup.js";
 import { getComments, postComment } from "./modules/commentsHandler.js";
-import { postLikes,getLikes } from "./modules/likesHandler";
-
-
-
+import { postLikes, getLikes } from "./modules/likesHandler.js";
 
 const popupHook = document.querySelector(".popup-hook");
 const itemContainer = document.querySelector(".cards");
 
 (async () => {
-  const meals = await getData();
+  const meals = await getAllMeal();
   const likes = await getLikes();
-  displayInHtml({meals,likes});
+  displayInHtml({ meals, likes });
 })();
 
 // Displaying popup
@@ -67,10 +65,11 @@ popupHook.addEventListener("click", (e) => {
   popupHook.innerHTML = "";
 });
 
-cardContainer.addEventListener('click',(e)=>{
+// handle likes
+cardContainer.addEventListener('click', (e) => {
   const closeLikeIcon = e.target.closest('.item-icon');
-  if(!closeLikeIcon) return;
+  if (!closeLikeIcon) return;
   closeLikeIcon.style.color = 'red';
-  const {id} = closeLikeIcon.dataset;
-  postLikes({item_id:id});
-})
+  const { id } = closeLikeIcon.dataset;
+  postLikes({ item_id: id });
+});
