+++ src/modules/likesHandler.js
@@ -1,23 +1,17 @@
-import { INVOLVEMENT_API_URL } from "../globals";
+import { INVOLVEMENT_API_URL } from "../globals.js";
 
-export const postLikes = async (data)=>{
-    await fetch(`${INVOLVEMENT_API_URL}/likes`,{
+export const postLikes = async (data) => {
+    await fetch(`${INVOLVEMENT_API_URL}/likes`, {
         method: "POST",
         body: JSON.stringify(data),
         headers: {
             'Content-type': 'application/json; charset=UTF-8',
           },
     });
-    
+};
 
-}
-
-export const getLikes = async ()=>{
+export const getLikes = async () => {
     const response = await fetch(`${INVOLVEMENT_API_URL}/likes`);
     const data = await response.json();
     return data;
-}
-
-
-
-
+};
