+++ src/style.css
@@ -115,7 +115,6 @@ h1 {
   font-weight: 600;
 }
 
-
 .card-btn {
   font-size: 2rem;
   padding: 1rem 1.5rem;
